requirements.txt
streamlit
pandas
requests
openai
PyPDF2
python-dotenv
openpyxl